[ revert ] 백버튼 쪽 컴포넌트 바깥으로 빼기
